+++ src/main/java/com/infinite/solution/phoneshop/controller/ModelController.java
@@ -30,7 +30,7 @@ public class ModelController {
 	}
 	
 	@GetMapping("{id}")
-	public ResponseEntity<?> getById(@PathVariable("id") Integer modelId){
+	public ResponseEntity<?> getById(@PathVariable("id") Long modelId){
 		Model model = modelService.getById(modelId);
 		return ResponseEntity.ok(modelMapper.toModelDto(model));
 	}
