+++ src/main/java/com/infinite/solution/phoneshop/exceptions/ResourceNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends ApiServiceException{
 
-	public ResourceNotFoundException(String entityName, Integer id) {
+	public ResourceNotFoundException(String entityName, Long id) {
 		super(HttpStatus.NOT_FOUND, String.format("%s_id = %d is not found", entityName,id));
 	}
