+++ src/main/java/com/infinite/solution/phoneshop/repository/ModelRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.infinite.solution.phoneshop.entity.Model;
 
-public interface ModelRepository extends JpaRepository<Model, Integer>{
-	List<Model> findByBrandId(Integer brandId);
+public interface ModelRepository extends JpaRepository<Model, Long>{
+	List<Model> findByBrandId(Long brandId);
 }
