+++ src/main/java/com/infinite/solution/phoneshop/service/ModelService.java
@@ -7,8 +7,8 @@ import com.infinite.solution.phoneshop.entity.Model;
 
 public interface ModelService {
 	Model create(Model model);
-	Model getById(Integer id);
+	Model getById(Long id);
 	//Model update(Integer id, ModelDTO modelDto);
 	//void remove(Integer id);
-	List<Model> getModelsByBrandId(Integer brandId);
+	List<Model> getModelsByBrandId(Long brandId);
 }
